LiveKit_AI Agent/requirements.txt
livekit-agents
livekit-plugins-google
livekit-plugins-noise-cancellation

python-dotenv~=1.1.1
requests~=2.32.5
psutil~=7.1.3

duckduckgo-search
langchain-community

h11~=0.16.0
pip~=22.3.1
attrs~=24.2.0
distro~=1.9.0
pillow~=12.0.0
utils~=1.0.1
tornado~=6.4.1
Jinja2~=3.1.2
sniffio~=1.3.1
fastapi~=0.125.0
pydantic~=2.11.7
starlette~=0.50.0
websockets~=15.0.1
anyio~=4.12.0
orjson~=3.11.5
click~=8.3.1
uvicorn~=0.38.0
watchfiles~=1.1.1
config~=0.5.1
rich~=14.2.0
itsdangerous~=2.2.0
PyYAML~=6.0.2
httpx~=0.28.1
setuptools~=65.5.0
packaging~=25.0
pyparsing~=3.2.3
zipp~=3.23.0
livekit~=1.0.20